HDDS-1597. Remove hdds-server-scm dependency from ozone-common. Contributed by Elek, Marton.
