+++ hadoop-hdds/framework/src/main/java/org/apache/hadoop/hdds/server/ServerUtils.java
@@ -203,4 +203,16 @@ public final class ServerUtils {
     conf.set(HddsConfigKeys.OZONE_METADATA_DIRS, path);
   }
 
+  public static File getDBPath(Configuration conf, String dbDirectory) {
+    final File dbDirPath =
+        getDirectoryFromConfig(conf, dbDirectory, "OM");
+    if (dbDirPath != null) {
+      return dbDirPath;
+    }
+
+    LOG.warn("{} is not configured. We recommend adding this setting. "
+            + "Falling back to {} instead.", dbDirectory,
+        HddsConfigKeys.OZONE_METADATA_DIRS);
+    return ServerUtils.getOzoneMetaDirPath(conf);
+  }
 }
